requirements.txt
docopt==0.6.1
exceptiongroup==1.2.2
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==7.4.4
semantic-version==2.3.0
smmap==5.0.1
SQLAlchemy==2.0.25
tomli==2.0.2
typing_extensions==4.12.2
Werkzeug==3.0.1
psycopg2
pyjwt==2.8.0
numpy==1.26.3
pandas==2.1.4
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
setuptools==69.0.3